+++ richat/src/richat/server.rs
@@ -30,6 +30,7 @@ impl RichatServer {
                     messages.clone(),
                     move || connections_inc.increment(1), // on_conn_new_cb
                     move || connections_dec.decrement(1), // on_conn_drop_cb
+                    VERSION,
                     shutdown.clone(),
                 )
                 .await?