+++ CHANGELOG.md
@@ -14,6 +14,8 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Features
 
+- proto: add version to quic response ([#115](https://github.com/lamports-dev/richat/pull/115))
+
 ### Breaking
 
 ## 2025-05-30